requirements.txt
baidu-aip==2.2.18.0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
numpy==1.19.5
opencv-contrib-python
pandas==0.25.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
synonyms==3.16.0
translate==3.6.1
py2neo==4.2.0

# -i https://pypi.tuna.tsinghua.edu.cn/simple
